symbolize_test: account for platforms with ".sdata" small data sections
A handful of ELF targets use a ".sdata" section for small data.  The
default maximum size of symbols in this section as well as the gcc flag
to change that maximum size differ between two targets.  On such
platforms, if there is no data large enough to exceed the small data
limit then the ".data" section will be excluded entirely.  This changes
the check to ensure that at least one of a ".sdata" section OR a ".data"
section is present.